+++ sql-infer-core/src/inference/datatypes.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use crate::{
     inference::{InformationSchema, SqlType, UseInformationSchema},
     parser::Column,
@@ -6,11 +8,20 @@ use crate::{
 pub struct TextLength;
 
 impl UseInformationSchema for TextLength {
-    fn apply(&self, schema: Option<&InformationSchema>, _: &Column, column: &mut super::QueryItem) {
+    fn apply(
+        &self,
+        schemas: &HashMap<Column, InformationSchema>,
+        column: &Column,
+        item: &mut super::QueryItem,
+    ) {
+        let schema = schemas.get(column);
         let Some(schema) = schema else {
             return;
         };
-        if let SqlType::Char { length } | SqlType::VarChar { length } = &mut column.sql_type {
+        if includes_cast(column) != Some(true) {
+            return;
+        }
+        if let SqlType::Char { length } | SqlType::VarChar { length } = &mut item.sql_type {
             if let Some(character_maximum_length) = schema.character_maximum_length {
                 *length = Some(character_maximum_length as u32)
             }
@@ -21,14 +32,23 @@ impl UseInformationSchema for TextLength {
 pub struct DecimalPrecision;
 
 impl UseInformationSchema for DecimalPrecision {
-    fn apply(&self, schema: Option<&InformationSchema>, _: &Column, column: &mut super::QueryItem) {
+    fn apply(
+        &self,
+        schemas: &HashMap<Column, InformationSchema>,
+        column: &Column,
+        item: &mut super::QueryItem,
+    ) {
+        let schema = schemas.get(column);
         let Some(schema) = schema else {
             return;
         };
+        if includes_cast(column) != Some(true) {
+            return;
+        }
         if let SqlType::Decimal {
             precision,
             precision_radix,
-        } = &mut column.sql_type
+        } = &mut item.sql_type
         {
             if let Some((numeric_precision, numeric_precision_radix)) =
                 schema.numeric_precision.zip(schema.numeric_precision_radix)
@@ -38,4 +58,17 @@ impl UseInformationSchema for DecimalPrecision {
             };
         }
     }
+}
+
+fn includes_cast(column: &Column) -> Option<bool> {
+    Some(match column {
+        Column::DependsOn { .. } => false,
+        Column::Maybe { column } => includes_cast(column)?,
+        Column::Either { left, right } => Option::zip(includes_cast(left), includes_cast(right))
+            .map(|(left, right)| left || right)?,
+        Column::Cast { .. } => true,
+        Column::BinaryOp { .. } => return None,
+        Column::Unknown { .. } => return None,
+        Column::Value { .. } => return None,
+    })
 }