+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# Unreleased
+
+## Breaking Changes
+- Empty queries cause errors
+
+## Added
+- `sql-infer-cli analyze` command  
+    It is now possible to run `sql-infer-cli analyze [columns|tables|columns-with-db] <queries>` on a list of comma separated paths or sql queries in order to see what the parser is doing.
+- Support for nested expressions (e.g. `(x)`) 
+- Partial support for casts
+- Support for values (e.g. `1`)
+- Support for certain binary operations: `<`, `<=`, `>`, `>=`, `=`, `!=`, `and`, `or`, `xor`, `+`, `-`, `*`, `/`, `%` and `||`.
+- Support for `count` by assuming it to be an integer value.
+
+## Internal changes
+- Information schema is now computed and cached for every `Column`.  
+
 # 0.10.0
 
 ## Added